schoolinfonepal-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.2
Pillow>=10.0.0
python-slugify>=8.0.1
